feat: 신고 기능 구현을 위한 엔티티 구현
